Insanely slow but at least partly working proof of concept
This is painfully slow but it does work to some extent; I can boot the
menu on the games SSD image in BeebEm with no sideways RAM writes
enabled. The games I've tried tend to not work, but I think this may be
due to my use of &70 as temporary workspace.